Crate sui_package_alt

Structs

BuildParams

PublishedMetadata

Note: Every field should be optional, and the system can pick sensible defaults (or error out) if fields are missing.

SuiFlavor

Functions

find_environment

Determine the correct environment to use for the package system based on

mainnet_environment

testnet_environment